Curve out the logic to bundler
